Revert "static_cast directly from double to enum"
This reverts commit b48b07fde6d92ac1347f3e25300d92d72d125b62 as directly
casting from double to enum is not supported by (any version of) MSVS,
even if it should be.

+++ gpt_input.cpp
@@ -404,7 +404,9 @@ void gpt_input::DoHarmonize()
 
     oenum_alb_or_anb const alb_anb =
         static_cast<oenum_alb_or_anb>
-            (database_->Query(DB_AgeLastOrNearest)
+            (static_cast<int>
+                (database_->Query(DB_AgeLastOrNearest)
+                )
             );
     DateOfBirth.minimum_and_maximum
         (minimum_birthdate(IssueAge.maximum(), EffectiveDate.value(), alb_anb)
@@ -548,7 +550,9 @@ void gpt_input::DoTransmogrify()
 
     oenum_alb_or_anb const alb_anb =
         static_cast<oenum_alb_or_anb>
-            (database_->Query(DB_AgeLastOrNearest)
+            (static_cast<int>
+                (database_->Query(DB_AgeLastOrNearest)
+                )
             );
 
     int apparent_age = attained_age
